Fix spelling errors
Great spelling and grammar are important parts of a quality product
as they can be indicators to the level of attention to detail used
for that product.

Small spelling mistakes can go unnoticed but can negatively impact
the opinion of the user even though the underlying product itself
may be excellent.