1. NPM Module and Bower component support. Updated package.json and gulpfile to allow publishing a NPM module.
Added a bower.json file to publish a Bower component.

Creating a NPM module basically consisted of adding some information in the package.json
(`main` - main JS file of the module - and `style` - main stylesheet - properties) and
precompiling everything into the dist/ folder. This way developers will have access
to the compiled code when installing the module in their project (`npm install
firebaseui`). The default gulp command will automatically generate the NPM module.

Bower packaging consists of creating the bower.json file. It contains metadata on the
component. Developers can then run `bower install firebaseui` and fetch the JS and CSS
files in their app by using paths to the dist folder of the bower components
(e.g. src="bower_components/dist/firebase-ui-auth.js")

Bower file specs:
https://github.com/bower/spec/blob/master/json.md

2. Travis and SauceLabs support.
Configured Travis to run tests on SauceLabs.

I've added a new command:
npm test
that alias npm run test. It is also the usual default command to run tests in a
NPM project.
See the sauce_browsers.json file for the list of browsers covered.

It sill supports running phantomJS by default, so developers can use it in their local
env.

How it works:
SauceConnect creates a tunnel to remote instances of SauceLabs browsers, which exposes
the localhost to the remote instance. This tunnel gets attributed an ID. Protractor
then uses this ID to connect to the distant browser and run tests. The HTTP server
that serves test files is run locally, and the tunnel is meant for the remote instances to
access it.
When using Travis, SauceConnect will be automatically started by the process (it's an
addon directly offered by Travis). The tunnel ID will be exposed as the
TRAVIS_JOB_NUMBER.

Travis configuration:
https://docs.travis-ci.com/user/languages/javascript-with-nodejs/
Notice I'm using the secure way of storing the access key, so it can safely be pushed
to the public GitHub repo.

SauceLabs configuration for Travis:
https://docs.travis-ci.com/user/sauce-connect/

Using protractor with SauceLabs:
https://support.saucelabs.com/customer/portal/articles/2029412-running-protractor-tests-on-sauce-labs
A lot of things happen under the hood here, like connecting to the remote browser
Webdriver instance through the tunnel.

PiperOrigin-RevId: 139406176
Change-Id: Ie5ca57218b3df5ccaee3fc187b8ac20ac31618fd