Fixes firebaseui.auth.idp.getAuthCredential credential initializer which was breaking account linking flow.
Added login_hint to Google OAuth flow when the email is specified.
This is applied in the following 2 cases:
1. Account linking is triggered and the user needs to sign in to existing Google account.
To force the user to sign in to the expected Google account, we pass the login hint with the email.
2. If a user selects email login and the Google account already exists for that email,
when clicking continue to sign in with Google account, we pass the login_hint for that account.